feat: Add create-topic functionality
This commit adds the functionality to create a new topic in the application. It includes changes to the create-topic component's HTML and TypeScript files, as well as the topics service. The user can now enter a title and description for the topic and submit it, which will add the new topic to the list of topics.